fix(ios): cast SessionUser.id (Int) a Int64 para APIs userPublic/userFollowingUsers

+++ Temazo/UI/MainScreen.swift
@@ -311,7 +311,7 @@ struct MainScreen: View {
                 },
                 onPublicProfileClick: {
                     guard let me = auth.currentUser else { return }
-                    detailStack.append(.userPublic(id: me.id, username: nil))
+                    detailStack.append(.userPublic(id: Int64(me.id), username: nil))
                 },
                 onRecapClick: { detailStack.append(.recap) },
                 onNotificationsClick: { detailStack.append(.notifications) }
